openadapt/strategies/cursor_replay_strategy.py
#cursor_replay_strategy.py
from .vanilla import VanillaReplayStrategy


class CursorReplayStrategy:
    pass


class CursorReplayStrategy:
    CursorReplayStrategy(VanillaReplayStrategy)
    def __int__(self,*args,**kwargs):
        super().__init__(*args,**kwargs)
        #initialize any additional attributes if needed


        def paint_red_dot(self,image,target_location):
            """Paint a red dot on the image at the specified target location.

            :param image: The image to paint on
            :param
            target_location: The (x,y) coordinates for the red dot
            :return Modified image with the red dot """

            #implement the logic to paint a red dot on the image

            pass
        def process(self,image):
            """

            Process the image to suggest target locations and paint red dots.

            :param image: The image to process
            :return: Modified image with suggested target locations marked
            """

            target_location = self.suggest_target_location(image)

            image_with_dot = self.paint_red_dot(image,target_location)

            return image_with_dot


        def suggest_target_location(self,image):
            """
            Suggest target location on the image.

            :param image: The image to analyze
            :return: The (x,y) coordinates of the suggested target location
            """

            #implement the logic to suggest target location

            pass


