tests/openadapt/test_cursor_replay_strategy.py
import unittest
from openadapt.strategies.cursor_replay_strategy import CursorReplayStrategy

class TestCursorReplayStrategy(unittest.TestCase):
    def setUp(self):
        self.strategy = CursorReplayStrategy()

        def test_paint_red_dot(self):
            # test the paint red dot method

            image = ...
            #Load or create a test image

            target_location = (50,50)
            result_image = self.strategy.paint_red_dot(image,target_location)
            #Add assertions to check if the dot is correctly painted

            self.assertIsNotNone(result_image)



            def test_process(self):
                self.assertIsNotNone(result_image)
        # Test the process method
        image = ...
        result_image = self.strategy.process(image)

        # add assertions to validate the processing


        if __name__ == '__main__':unittest.main()